Time: 3 ms (94.03%) | Memory: 61.7 MB (61.45%) - LeetSync

+++ 2432-number-of-zero-filled-subarrays/number-of-zero-filled-subarrays.java
@@ -1,7 +1,7 @@
 class Solution {
     public long zeroFilledSubarray(int[] nums) {
         long result = 0;
-        long temp = 0;
+        int temp = 0;
         for(int i = 0; i < nums.length; i++){
             if(nums[i] == 0){
                 temp++;
